requirements.txt
absl-py==0.7.1
astor==0.8.0
audioread==2.1.7
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.0.2
gast==0.2.2
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-core==1.0.3
google-cloud-storage==1.22.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.20.0
grpcio-tools==1.20.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
librosa==0.6.3
llvmlite==0.28.0
locustio==0.11.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
msgpack==0.6.1
numba==0.43.1
numpy==1.16.3
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pydub==0.23.1
PyMySQL==0.9.3
pytz==2019.1
pyzmq==18.0.1
requests==2.21.0
resampy==0.2.1
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sox==1.3.7
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-serving-api==1.13.0
termcolor==1.1.0
urllib3==1.24.2
Werkzeug==0.15.2
